Fix so MFC\ole\oleview is able to compile in VS2017. Fix to compile for Unicode APIs. Fix crash bug when a function has more or equal count of parameters than MAX_NAMES. It should now properly cut off without a crash. Add support for up to 128 parameters. Add support for ITypeInfo. Let CTypeLibTreeView select first item on update, so that other tools can extract the complete IDL source without the need to do special operations with the automation object.

+++ VC2010Samples/MFC/ole/oleview/iviewers/idataobj.cpp
@@ -511,7 +511,7 @@ HRESULT CIDataObjectDlg::GotData( LPFORMATETC lpfetc, LPSTGMEDIUM lpstm )
 		#else
 			if (lpfetc->cfFormat == CF_TEXT)
 			{
-				UINT n = _mbstrlen( (LPCSTR)lpsz ) ;
+				UINT n = (UINT)_mbstrlen( (LPCSTR)lpsz ) ;
 				TCHAR* sztemp = new TCHAR[n+1] ;
 				mbstowcs( sztemp, (LPSTR)lpsz, n ) ;
 				WriteToOutput( sztemp ) ;
